Renaming to "spline".
Also includes forgotten files from the previous commit; oops.

"Splinter" has been claimed as a different project since I started
working on this, and "spline" is the original name I liked but already
used for an earlier and crappier approach.  But the old "spline" never
made it to pypi (and was never used by anyone but me anyway), so I might
as well reclaim it for this hopefully-better attempt.